conditional split-view render
